Cформировал начало для веба

+++ src/main/java/ru/IS_122/AutomationOfPharmacyChainOperations/model/UserOfPharmacy.java
@@ -3,6 +3,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -17,7 +18,7 @@ public class UserOfPharmacy {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "allobjects_seq")
     @SequenceGenerator(name = "allobjects_seq", sequenceName = "allobjects", allocationSize = 1)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private BigDecimal id;
 
     @Column(name = "login", nullable = false, length = 50)
     @NonNull
@@ -46,6 +47,6 @@ public class UserOfPharmacy {
 
     @Column(name = "role", nullable = false, length = 0)
     @NonNull
-    private Long role;
+    private BigDecimal role;
 
 }
